Исправление недочетов 4-ого задания
