+++ js/generate-date.js
@@ -1,5 +1,21 @@
-import { names, messages } from './arrays.js';
-import { chooseNumber} from './util.js';
+import { chooseNumber } from './util.js';
+
+// Коллекция имён (массив)
+const names = [
+  'Алексей', 'Александр', 'Андрей', 'Александра', 'Ангелина', 'Борис', 'Бротислав', 'Бэн', 'Бонни', 'Бэлла',
+  'Владимир', 'Вячеслав', 'Воланд', 'Виктория', 'Валерия', 'Григорий', 'Герман', 'Геннадий', 'Галина', 'Гертруда',
+  'Дмитрий', 'Денис', 'Довлат', 'Дарья', 'Дана',
+];
+
+//Коллекция сообщений (массив)
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
 
 const generateMessage = () => {
   const index = chooseNumber(0, 5);
@@ -50,5 +66,4 @@ const generateFinal = (quantity) => {
   return photos;
 };
 
-generateFinal(25);
-
+export { generateFinal };
